% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-collection.R
\name{aoi}
\alias{aoi}
\title{Access AOI}
\usage{
aoi(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An AOI object.
}
\description{
Returns the area of interest associated with an object.
}
