% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calc_nbr}
\alias{calc_nbr}
\title{Calculate the Normalized Burn Ratio}
\usage{
calc_nbr(x)
}
\arguments{
\item{x}{A composite SpatRaster containing
"nir08" and "swir22".}
}
\value{
A single-layer SpatRaster.
}
\description{
Calculate the Normalized Burn Ratio
}
