% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{classify_burn_severity}
\alias{classify_burn_severity}
\title{Classify burn severity from dNBR}
\usage{
classify_burn_severity(x)
}
\arguments{
\item{x}{A single-layer dNBR SpatRaster.}
}
\value{
A categorical SpatRaster containing burn severity classes.
}
\description{
Classifies a dNBR raster into burn severity classes.
}
