% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{create_aoi}
\alias{create_aoi}
\title{Create an Area of Interest}
\usage{
create_aoi(
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  lon = NULL,
  lat = NULL,
  radius = NULL,
  crs = "EPSG:4326"
)
}
\arguments{
\item{xmin, xmax, ymin, ymax}{Bounding box coordinates.}

\item{lon, lat}{Longitude and latitude of the centre point.}

\item{radius}{Radius (metres) around the centre point.}

\item{crs}{Coordinate reference system of the input coordinates.}
}
\value{
An sbr_aoi object.
}
\description{
Create an Area of Interest (AOI) from either a bounding box or
a point and radius.
}
\examples{
\dontrun{
aoi <- create_aoi(
    xmin = 1.61,
    ymin = 52.24,
    xmax = 1.64,
    ymax = 52.26
)
}
}
