% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rgb}
\alias{plot_rgb}
\title{Plot differenced Normalized Burn Ratio}
\usage{
plot_rgb(x, title = "RGB composite", subtitle = NULL)
}
\arguments{
\item{x}{A single-layer dNBR SpatRaster.}

\item{title}{Plot title.}

\item{subtitle}{Plot subtitle}
}
\value{
A ggplot object.
}
\description{
Plot differenced Normalized Burn Ratio
}
