% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_s2}
\alias{search_s2}
\title{Search Sentinel-2 imagery}
\usage{
search_s2(aoi, start, end)
}
\arguments{
\item{aoi}{An sbr_aoi object.}

\item{start}{Start date.}

\item{end}{End date.}
}
\value{
A raw STAC result.
}
\description{
Search Sentinel-2 imagery
}
